sharehw/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
Pillow==10.0.0
redis==5.0.0
bcrypt==4.0.1
python-magic==0.4.27

# Testing dependencies
pytest==7.4.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Development tools
black==23.9.1
flake8==6.1.0
mypy==1.5.1
